fix filter by price

+++ src/main/java/pe/edu/upc/service/BoxService.java
@@ -11,8 +11,8 @@ public interface BoxService {
 	List<BoxDTO> listByPersonalized(boolean personalized);
 	List<BoxDTO> listByClientsUserLoginEmail(String email);
 	List<BoxDTO> listByPersonalizedAndName(boolean personalized, String name);
-	List<BoxDTO> listByPersonalizedAndPriceGreaterThanEqualAndPriceLessThanEqual(boolean personalized, float priceMin, float priceMax);
-	List<BoxDTO> listByPersonalizedAndNameAndPriceGreaterThanEqualAndPriceLessThanEqual(boolean personalized, String name, float priceMin, float priceMax);
+	List<BoxDTO> listByPersonalizedAndPriceGreaterThanEqualAndPriceLessThanEqual(boolean personalized, double priceMin, double priceMax);
+	List<BoxDTO> listByPersonalizedAndNameAndPriceGreaterThanEqualAndPriceLessThanEqual(boolean personalized, String name, double priceMin, double priceMax);
 	BoxDTO createPersonalized(BoxCreateDTO boxCreateDTO);
 	BoxDTO createDefault(BoxCreateDTO boxCreateDTO);
 	BoxDTO saveDefaultBox(SaveDefaultBoxDTO saveDefaultBoxDTO);
